ci(terraform-fmt): add format check for TF files
Add terraform files' formatting check on each PR and push to main.

ci(terraform fmt): force tf fmt check to run

Force the TF check to run even when there are merge conflicts.

ci(terraform fmt): fix the name of the main branch

Fix the name of the main branch given, to make the check run correctly
on PR to main.

Fix the check not triggering.